requirements.txt
pymongo==4.6.1
sentence-transformers>=2.7.0
faiss-cpu>=1.12.0
numpy==1.26.3
openai>=1.10.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
flask>=3.0.0
flask-cors>=4.0.0
